Changed Parameters type to be object rather than JObject
This avoid implementing our own JsonSchema as the user will be able to provide their own. If the user wants to use a schema like Newtonsoft, NJsonSchema, or their own implementation it can handle it, if the user passes in a dictionary<string, obj> it can handle it, or if the user passes in a string formatted as JSON it can handle it.