Refactored non-streaming chat endpoints
Added a helper function which will allow us to get different types of results from the returned Message.

Updated GetResponseFromChatbotAsync to use this method.

Added
GetFuntion_CallResponseAsync which returns the function_call from the response message.

This will allow us to be a bit more modular in the future if more message type responses get added.